restructured core functions and main file
