--- src/orphan_detection_steps.py
@@ -1,3 +1,0 @@
-"""This file contains all function steps to detect orphan pages for a single domain
-except the ones for the dynamic url detection step"""
-
